Try generating a local prisma client for tests
