+++ packages/uploads/.gitignore
@@ -1,3 +1,4 @@
 src/__tests__/migrations/*
 src/__tests__/for_unit_test.db*
 .attw.json
+src/__tests__/prisma-client/*
