+++ packages/uploads/src/__tests__/unit-test-schema.prisma
@@ -5,6 +5,7 @@ datasource db {
 
 generator client {
   provider = "prisma-client-js"
+  output   = "./prisma-client" // <-- we generated a local prisma client so it doesn't interfere with the mono repo
 }
 
 model Dummy {
